# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Medshelf API",
        "description": "Documentacion OpenAPI para los endpoints de Medshelf.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Base URL"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List products",
                "operationId": "c1ff6f862214e1896c59cfe0491ce0e8",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "description": "Filter by product name (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "required": [
                                                "items",
                                                "nextCursor"
                                            ],
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/ProductResponse"
                                                    }
                                                },
                                                "nextCursor": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "required": [
                                                "items",
                                                "totalCount",
                                                "page",
                                                "size",
                                                "hasMorePages"
                                            ],
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/ProductResponse"
                                                    }
                                                },
                                                "totalCount": {
                                                    "type": "integer",
                                                    "minimum": 0
                                                },
                                                "page": {
                                                    "type": "integer",
                                                    "minimum": 1
                                                },
                                                "size": {
                                                    "type": "integer",
                                                    "maximum": 100,
                                                    "minimum": 1
                                                },
                                                "hasMorePages": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Add product",
                "operationId": "399770d33fa4672f1216707d0e236763",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "pharmaceuticalForm",
                                    "composition"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Paracetamol"
                                    },
                                    "netContent": {
                                        "properties": {
                                            "value": {
                                                "type": "number",
                                                "minimum": 0,
                                                "example": 100
                                            },
                                            "unit": {
                                                "type": "string",
                                                "example": "ml"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "totalQuantity": {
                                        "type": "number",
                                        "minimum": 0,
                                        "example": 20
                                    },
                                    "pharmaceuticalForm": {
                                        "type": "string",
                                        "example": "Jarabe"
                                    },
                                    "composition": {
                                        "properties": {
                                            "referenceAmount": {
                                                "type": "number",
                                                "minimum": 0,
                                                "example": 5
                                            },
                                            "activeIngredients": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Acetaminofen"
                                                        },
                                                        "strength": {
                                                            "properties": {
                                                                "value": {
                                                                    "type": "number",
                                                                    "minimum": 0,
                                                                    "example": 500
                                                                },
                                                                "unit": {
                                                                    "type": "string",
                                                                    "example": "mg"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/{productId}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product details",
                "operationId": "59868682c75d47fe2401d30a6e9d88e8",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ProductResponse": {
                "required": [
                    "id",
                    "name",
                    "pharmaceuticalForm",
                    "createdAt",
                    "composition"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "netContent": {
                        "properties": {
                            "value": {
                                "type": "number"
                            },
                            "unit": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "totalQuantity": {
                        "type": "number",
                        "nullable": true
                    },
                    "pharmaceuticalForm": {
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "consumptionType": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "composition": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}